+++ lib/vph_web_date_picker.dart
@@ -1,4 +1,3 @@
 library;
 
 export 'src/web_date_picker.dart';
-export 'src/web_month_picker.dart';